requirements.txt
# requirements.txt
python-telegram-bot
fastapi
uvicorn
requests  # You can remove this if you want
httpx     # <-- Add this line
python-dotenv